requirements.txt
python-dotenv==1.0.0
pandas==2.1.4
sqlalchemy==2.0.23
streamlit==1.29.0
google-generativeai==0.3.2
